rename files and restructure packages
