+++ internal/controller/discovery/core/loader.go
