+++ internal/controller/discovery/core/message.go
