+++ internal/controller/discovery/loaders/factory.go
