+++ internal/controller/discovery/loaders/http/loader.go
