+++ internal/controller/discovery/loaders/http/loader_test.go
