+++ internal/controller/discovery/reconciler/message_processor.go
@@ -1,4 +1,4 @@
-package discovery
+package reconciler
 
 import (
 	"context"
